add ui to show cart amount info for checkout

+++ packages/custom/shopCatalog/public/controllers/checkoutController.js
@@ -5,13 +5,14 @@
 	function($scope, Global, cartService) {
 	  $scope.global = Global;
 	  $scope.items = [];
+        $scope.shipping = 10;
+        $scope.tax = 0;
 
 	  cartService.getCart()
 		.$promise
 		.then(function(cart){
 		  $scope.items = cart.items;
 		});
-
 	  $scope.increaseQuantity = function(item) {
           item.quantity+= 1;
           cartService.addToCart(item.product, item.quantity);
@@ -26,7 +27,9 @@
 	  };
 
 	  $scope.removeFromCart = function(product) {
-		cartService.removeFromCart(product);
+          if(confirm('Are you sure you want to delete this item ?')) {
+              cartService.removeFromCart(product);
+          }
 	  };
 	}
   ]);
